app/requirements.txt
python-dotenv==1.0.0 
streamlit==1.40.*
requests==2.32.*
mysql-connector-python
neo4j